Activities structure finally working

+++ Application/src/main/java/com/example/android/networkconnect/model/Position.java
@@ -7,10 +7,10 @@ import org.json.JSONObject;
  * Created by mateu_000 on 2015-04-26.
  */
 public class Position {
-    public float Latitude;
-    public float Longitude;
+    public double Latitude;
+    public double Longitude;
 
-    public Position(float longitude, float latitude) {
+    public Position(double longitude, double latitude) {
         Longitude = longitude;
         Latitude = latitude;
     }
